requirements.txt
asgiref==3.8.1
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
Django==5.1.7
django-cors-headers==4.7.0
django-environ==0.12.0
django-phonenumber-field==8.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
environ==1.0
gunicorn==23.0.0
idna==3.10
packaging==25.0
paho-mqtt==2.1.0
phonenumbers==9.0.9
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
requests==2.32.4
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
